Moves A11y explorer into separate file.
